\name{adt-package}
\alias{adt-package}
\alias{adt}
\docType{package}
\title{
\packageTitle{adt}
}
\description{
\packageDescription{adt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{'adt' is one R package which can be used to compare the similarity of two data sets.}
\packageIndices{Our method is based on projection, and it starts with the results of PCA (Principal Component Analysis}
~~ functions ~~
}
\author{
\packageAuthor{adt}

Maintainer: \packageMaintainer{adt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
